requirements.txt
torch
numpy
matplotlib
Pillow
